Поправил схемы БД

+++ src/main/java/org/wolrus/digital_system/repository/GroupRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.wolrus.digital_system.entity.GroupEntity;
 
 
-public interface GroupRepository extends JpaRepository<GroupEntity, Long> {
+public interface GroupRepository extends JpaRepository<GroupEntity, Integer> {
     Integer countAllByIsOpenIsTrueAndAge(String type);
 }
